feat(definitions): added the TextRange struct
The TextRange struct defines a start and end location for keeping track
of text spans inside of a long text.

Also added the "calculate/2" function which returns a new TextRange
where the start location is set to the given start location and the end
location to the start location incremented by the length of the given
text.

Also added the test for the calculate/2 function.