feat(lexer): added the Lexer struct
The Lexer is used for tokenizing file contents of a source file to more
general words (lexemes).

For now it can lex the following things:
- keywords (by the given keywords list)
- identifiers
- strings
- characters (char's)
- numbers (integer + float + underscored (for reading purposes))
- single line comments (starting with "//")
- multi line comments (starting with "/*" and ending with "*/")

The lexer can throw a "UnexpectedEndOfFileError" when a string, char or
multi line comment is not properly closed before the end of the file was
reached.

Also added the test file for the lexer.